litreview/requirements.txt
appnope==0.1.3
asgiref==3.6.0
asttokens==2.2.1
autopep8==2.0.2
backcall==0.2.0
backports.zoneinfo==0.2.1
decorator==5.1.1
Django==4.2.1
executing==1.2.0
ipdb==0.13.13
ipython==8.12.2
jedi==0.18.2
matplotlib-inline==0.1.6
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
prompt-toolkit==3.0.39
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.10.0
Pygments==2.15.1
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.2
tomli==2.0.1
traitlets==5.9.0
typing_extensions==4.7.1
wcwidth==0.2.6
